[PRMT-2640] reverted most changes related to object locking defaulting back to -ehr-repo-bucket bucket
We found out that to implement object lock appropriately changes are required also at the client side
(ehr-transfer-service using pre-signed urls in this instance) so the story might end up way bigger than
we initially anticipated.

We still think objeck locking is an feature that we might need, therefore we'll introduce it separately
in a future story dedicated only to that.